[N-02] Removing unnecessary cast in Claim.sol
Audit Feedback:
>Within the `Claim` contract, the `uint256(amount)` cast is unnecessary.
>
>To improve the overall clarity, intent, and readability of the codebase, consider removing unnecessary casts.

+++ contracts/governance/Claim.sol
@@ -186,7 +186,7 @@ contract Claim is IClaimable, Ownable2Step {
         // If repurposing this contract, update to suit your needs.
         token.transfer(_destination, amount);
 
-        emit FundsRescued(_destination, uint256(amount));
+        emit FundsRescued(_destination, amount);
     }
 
     /**
